requirements.txt
altgraph==0.17.2
attrs==21.4.0
auto-py-to-exe==2.19.0
autopep8==1.6.0
beautifultable==1.1.0
black==22.3.0
bottle==0.12.19
bottle-websocket==0.2.9
click==8.1.3
dearpygui==1.6.1
Eel==0.12.4
filelock==3.6.0
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
iniconfig==1.1.1
iso8601==1.0.2
mypy-extensions==0.4.3
numpy==1.22.4
packaging==21.3
pandas==1.4.2
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyinstaller==5.0.1
pyinstaller-hooks-contrib==2022.4
pyparsing==3.0.9
pyserial==3.5
pysondb==1.6.7
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
scapy==2.4.5
six==1.16.0
toml==0.10.2
tomli==2.0.1
wcwidth==0.2.5
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.4.0
